requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
jinja2>=3.1.2
python-multipart>=0.0.6
itsdangerous>=2.1.2
bcrypt>=4.0.0
aiofiles>=23.2.1
starlette>=0.27.0
openpyxl>=3.1.0
ldap3>=2.9.1
python-dateutil>=2.8.2
requests>=2.31.0
